audio drivers: enable Werror
Part of #9460

Tested on both x86-64 and x86 builds

* If there is an operation between signed and unsigned int, the unsigned is changed to signed one,
  except for some operations (if before the operation the signed one is checked to be less than
  zero or not, then the signed one get cast to be unsigned (errors due to sign error when casting
  are impossible))
* When doing operations on pointers integers get cast to uintptr_t type
* Some functions expected int32 as one of the parameters however every call for the function uses
  uint32, so the input of is changed to uint32
* Deleted unused functions

+++ src/add-ons/kernel/drivers/audio/ac97/auich/multi.c
@@ -182,7 +182,7 @@ auich_ac97_set_mix(void *card, const void *cookie, int32 type, float *values) {
 
 
 static int32
-auich_create_group_control(multi_dev *multi, int32 *index, int32 parent,
+auich_create_group_control(multi_dev *multi, uint32 *index, int32 parent,
 	int32 string, const char* name) {
 	int32 i = *index;
 	(*index)++;
@@ -201,7 +201,7 @@ auich_create_group_control(multi_dev *multi, int32 *index, int32 parent,
 static status_t
 auich_create_controls_list(multi_dev *multi)
 {
-	uint32 	i = 0, index = 0, count, id, parent, parent2, parent3;
+	uint32 	i = 0,index=0, count, id, parent, parent2, parent3;
 	const ac97_source_info *info;
 
 	/* AC97 Mixer */
@@ -397,7 +397,7 @@ auich_get_mix(auich_dev *card, multi_mix_value_info * mmvi)
 	multi_mixer_control *control = NULL;
 	for (i = 0; i < mmvi->item_count; i++) {
 		id = mmvi->values[i].id - EMU_MULTI_CONTROL_FIRSTID;
-		if (id < 0 || id >= card->multi.control_count) {
+		if (id < 0 || (uint32)id >= card->multi.control_count) {
 			PRINT(("auich_get_mix : "
 				"invalid control id requested : %" B_PRIi32 "\n", id));
 			continue;
@@ -438,7 +438,7 @@ auich_set_mix(auich_dev *card, multi_mix_value_info * mmvi)
 	multi_mixer_control *control = NULL;
 	for (i = 0; i < mmvi->item_count; i++) {
 		id = mmvi->values[i].id - EMU_MULTI_CONTROL_FIRSTID;
-		if (id < 0 || id >= card->multi.control_count) {
+		if (id < 0 || (uint32)id >= card->multi.control_count) {
 			PRINT(("auich_set_mix : "
 				"invalid control id requested : %" B_PRIi32 "\n", id));
 			continue;
@@ -449,7 +449,7 @@ auich_set_mix(auich_dev *card, multi_mix_value_info * mmvi)
 			multi_mixer_control *control2 = NULL;
 			if (i+1<mmvi->item_count) {
 				id = mmvi->values[i + 1].id - EMU_MULTI_CONTROL_FIRSTID;
-				if (id < 0 || id >= card->multi.control_count) {
+				if (id < 0 || (uint32)id >= card->multi.control_count) {
 					PRINT(("auich_set_mix : "
 						"invalid control id requested : %" B_PRIi32 "\n", id));
 				} else {
@@ -509,7 +509,7 @@ auich_list_mix_controls(auich_dev *card, multi_mix_control_info * mmci)
 
 	if (auich_create_controls_list(&card->multi) < B_OK)
 		return B_ERROR;
-	for (i = 0; i < card->multi.control_count; i++) {
+	for (i = 0; (uint32)i < card->multi.control_count; i++) {
 		mmc[i] = card->multi.controls[i].mix_control;
 	}
 
@@ -582,7 +582,7 @@ auich_create_channels_list(multi_dev *multi)
 	chans = multi->chans;
 	index = 0;
 
-	for (mode=AUICH_USE_PLAY; mode!=-1;
+	for (mode=AUICH_USE_PLAY; (int32) mode!=-1;
 		mode = (mode == AUICH_USE_PLAY) ? AUICH_USE_RECORD : -1) {
 		LIST_FOREACH(stream, &((auich_dev*)multi->card)->streams, next) {
 			if ((stream->use & mode) == 0)
@@ -690,7 +690,7 @@ auich_get_description(auich_dev *card, multi_description *data)
 	// channels and input bus channels and finally auxillary channels,
 
 	LOG(("request_channel_count = %d\n",data->request_channel_count));
-	if (data->request_channel_count >= size) {
+	if (data->request_channel_count >= (int32)size) {
 		LOG(("copying data\n"));
 		memcpy(data->channels, card->multi.chans, size * sizeof(card->multi.chans[0]));
 	}
